requirements.txt
--extra-index-url=https://databricks-dash.aws.plotly.host/packages
dash-design-kit==1.8.1
dash==2.11.0
gunicorn==20.0.4
pandas>=1.1.5
dash-mantine-components==0.12.1
dash-iconify==0.1.2
sqlalchemy==1.4.44
dash-ag-grid==2.0.0
databricks-sql-connector==2.5.2
flask-sqlalchemy==3.0.3
websockets==11.0.3
plotly-express==0.4.1
databricks-cli==0.17.7
configparser
dash-chart-editor
dash-bootstrap-components
python-dotenv==1.0.0
pyfunctional>=1.4.3
dash-bootstrap-components>=1.2.0
dash-extensions==0.1.13
ipywidgets>=7.7.0
memory-profiler>=0.60.0
line-profiler>=3.5.1
pyarrow>=6.0.0
kaleido>=0.2.1
flask-cors>=3.0.10
trace-updater==0.0.9.1
plotly-resampler==0.9.1
polars==0.19.3
polars_plotly_resampler.tar.gz